### 2021-02-08 #3019 @whedon
[PRE REVIEW]: Precise calculations on hypercomplex numbers in C++
**Submitting author:** @AngryMaciek (<a href="http://orcid.org/0000-0003-1361-7301">Maciej Bak</a>)
**Repository:** <a href="https://github.com/AngryMaciek/hypercomplex" target ="_blank">https://github.com/AngryMaciek/hypercomplex</a>
**Version:** v1.0.0
