### 2021-02-14 #3019 @AngryMaciek
Dear @melissawm,

Thank you for accepting the editor's position here.

I do not have anyone specific for the reviewers in my mind but I would suggest one mathematician to double-check the algebra and one C++ expert to double-check the template programming and code structure.

Kind Regards,
Maciek
