### 2021-02-15 #3019 @melissawm
Hello, @AngryMaciek! Before we start, I'd like to suggest some minor changes to the paper:

- It would be nice to add the name of your software package in the title of your paper, and remove "The following repository.." As much as this is about a software repository, people reading your software paper may need a bit more context :)
- For the references [Hořeňovský,  M. (2018)], [The Boost Organization (1999)] and [van Heesch (2013)], it might be interesting to cite a specific version number.
- For other github repositories, I'm not sure how the "publishing date" was chosen. Can you clarify that? For example, for the [fed36 (2015)] reference, there appear to be changes in 2016. In cases where there's no release or version tag in the repository, it might be useful to either cite a specific commit or add to the citation the date where the last access was made. See, for example, [this document on citation principles for software](https://www.force11.org/software-citation-principles ).

Meanwhile I'll get started on finding reviewers. Thanks!
