+++ paper.bib
@@ -46,6 +46,7 @@ author = {ferd36},
 author = {{{The Boost Organization}}},
   title = {Boost C++ libraries},
   year = {1999},
+  version = {1.75.0},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/boostorg/boost}
