+++ paper.bib
@@ -45,7 +45,7 @@ author = {ferd36},
 @misc{boost,
 author = {{{The Boost Organization}}},
   title = {Boost C++ libraries},
-  year = {1999},
+  year = {2020},
   version = {1.75.0},
   publisher = {GitHub},
   journal = {GitHub repository},
