+++ paper.bib
@@ -55,7 +55,8 @@ author = {{{The Boost Organization}}},
 @misc{doxygen,
 author = {{{van Heesch, D.}}},
   title = {Doxygen},
-  year = {2013},
+  year = {2021},
+  version = {1.9.1},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/doxygen/doxygen}
@@ -73,7 +74,8 @@ author = {{{Google Inc.}}},
 @misc{catch2,
 author = {M. Hořeňovský},
   title = {Catch2},
-  year = {2018},
+  year = {2020},
+  version = {2.13.2},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/catchorg/Catch2}
