+++ paper.bib
@@ -36,7 +36,8 @@ author = {T. Hoppe},
 @misc{quaternions,
 author = {ferd36},
   title = {Quaternions},
-  year = {2015},
+  year = {2016},
+  commit = {8fc6c7},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/ferd36/quaternions}
