+++ paper.bib
@@ -37,7 +37,7 @@ author = {T. Hoppe},
 author = {ferd36},
   title = {Quaternions},
   year = {2016},
-  commit = {8fc6c7},
+  date = {15.02.2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/ferd36/quaternions}
