+++ paper.bib
@@ -36,8 +36,7 @@ author = {T. Hoppe},
 @misc{quaternions,
 author = {ferd36},
   title = {Quaternions},
-  year = {2016},
-  date = {15.02.2021},
+  year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/ferd36/quaternions}
