+++ paper.bib
@@ -65,7 +65,7 @@ author = {{{van Heesch, D.}}},
 @misc{cpplint,
 author = {{{Google Inc.}}},
   title = {cpplint},
-  year = {2009},
+  year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/google/styleguide}
