+++ paper.bib
@@ -27,7 +27,7 @@
 @misc{cd,
 author = {T. Hoppe},
   title = {Cayley Dickson},
-  year = {2015},
+  year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/thoppe/Cayley-Dickson}
