### 2021-02-16 2f7db6f
docs: minor changes in the wrap-up (#24)
