+++ paper.bib
@@ -27,7 +27,7 @@
 @misc{cd,
 author = {T. Hoppe},
   title = {Cayley Dickson},
-  year = {2015},
+  year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/thoppe/Cayley-Dickson}
@@ -36,7 +36,7 @@ author = {T. Hoppe},
 @misc{quaternions,
 author = {ferd36},
   title = {Quaternions},
-  year = {2015},
+  year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/ferd36/quaternions}
@@ -45,7 +45,8 @@ author = {ferd36},
 @misc{boost,
 author = {{{The Boost Organization}}},
   title = {Boost C++ libraries},
-  year = {1999},
+  year = {2020},
+  version = {1.75.0},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/boostorg/boost}
@@ -54,7 +55,8 @@ author = {{{The Boost Organization}}},
 @misc{doxygen,
 author = {{{van Heesch, D.}}},
   title = {Doxygen},
-  year = {2013},
+  year = {2021},
+  version = {1.9.1},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/doxygen/doxygen}
@@ -63,7 +65,7 @@ author = {{{van Heesch, D.}}},
 @misc{cpplint,
 author = {{{Google Inc.}}},
   title = {cpplint},
-  year = {2009},
+  year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/google/styleguide}
@@ -72,7 +74,8 @@ author = {{{Google Inc.}}},
 @misc{catch2,
 author = {M. Hořeňovský},
   title = {Catch2},
-  year = {2018},
+  year = {2020},
+  version = {2.13.2},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/catchorg/Catch2}
