### 2021-02-16 #3019 @AngryMaciek
Dear @melissawm,

* OK, changes included.
* OK, I have specified the most recent stable versions and adjuted the 'year' field respectively.
* I am generally following the example: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography These other three github repositories do not tag versions and the parsing engine for the paper does not support pointing to specific commits therefore I have modified the 'year' field for them to be 'last accessed' field. I believe that previously I have chosen the initial year of the project start, which does not make much sense as for a reference... It looks much better now, thanks!
