### 2021-03-10 AngryMaciek/hypercomplex#25 @dvberkel
Could you provide an example of how to use hypercomplex to solve "real world analysis problems"
Hi @AngryMaciek

I am missing a "real world" problem and how to solve it with `hypercomplex` in the documentation

**Describe the solution you'd like**
I would like to see a description of a problem that is well suited to solve with hypercomplex numbers, and see example code how to solve it with your software.

I have not considered alternatives yet, but I am open to suggestions.

**Additional context**
I am a reviewer for your submission to the Journal of Open Source Software. Right of the bat, I really like your project, the energy you devoted to the project, it's documentation and the state of the repository.

One of the items that I need to review, as per issue [openjournals/joss-reviews#3096](https://github.com/openjournals/joss-reviews/issues/3096), is the example usage. For starters, the available documentation perfectly describes how to use your work. But what I am missing is an "real world" problem being solved with your software.

That is why I opened this issue to see if we can come up with a "real world" problem that demonstrates the power of the software.
