### 2021-03-12 AngryMaciek/hypercomplex#25 @AngryMaciek
Hi @dvberkel,

Thanks for the kind words! Let's consider our options:

I don't think humanity at its current state of knowledge has any **real world** applications to  32>=dim. algebras.
Hopefully my package could contribute to a breakthrough... 😅

As for Sedonions, the only single resource I found was an ML publication from 2020 about their use in a neural network:
https://ieeexplore.ieee.org/document/9160921
This looks really cool and hopefully we might see some real progress in that direction in the future, but I think it's too early for me to implement any examples on my own. Plus, such a program would be rather complex and by itself could serve as another project...

OK, so we try Octonions. I was looking through the wiki and found some references to their applications:
https://en.wikipedia.org/wiki/Octonion#Applications
The problem here is that the applications they describe here are mostly in the field of physics, in which I have no training at all. Maybe if I was a physics PhD I could write some code that would perform meaningful calculations on octonions and provide a short interpretation but unfortunately I lack domain-specific knowledge. Apart from that they also mention robotics and (again) deep learning - but also rather specific tasks, way above my competence level.

Fine, let's focus on Quaternions. Here we can easily find that they represent rotations in 3D:
https://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation
I guess that I could implement a simple program that rotates an object in space. But this will still be an abstract proof of principle, not really a real world application.

I designed this small library mostly for these above-mentioned experts who work on Octonions and Sedonions and use them in their domain. I would like to aid all these ML specialists, physicists and robotic engineers solve their problems more efficiently. But I'm not one of them myself... Therefore I would claim that in this specific case it is hard to provide a real-world application. I hope you notice that from the nature of this library. It's not really aimed to solve one well-defined task. Please note that in your checklist it says: _Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)._ So I guess this submission might not be "ideal"...
