### 2021-03-15 AngryMaciek/hypercomplex#25 @dvberkel
Hi @AngryMaciek

I totally agree with your assessment of the "real world" applications. I searched around myself, and I could not easily find more compelling application then rotations.

I think you did your best effort
