### 2021-03-15 AngryMaciek/hypercomplex#25 @AngryMaciek
Thank you very much for your help and understanding.
