### 2021-03-19 #3096 @bonh
Hi @AngryMaciek !

Great software and super clean engineering and project management! I have a few comments, suggestions and so on. I will start to open issues in your repo so that we can discuss them.
