### 2021-03-19 AngryMaciek/hypercomplex#26 @bonh
JOSS: Substantial scholarly effort?
As this questions is crucial for publication in JOSS I will try to make my reasoning as transparent as possible!

For this question I am walking through the guidelines provided by [JOSS](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

1. _As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual._
The repo practically nails that requirement, right? (positive).
2. _Age of software (is this a well-established software project) / length of commit history._
The repo is very young (less 6 months, right?), without any forks (negative).
3. _Number of commits._
The number of commits is low (less 500, right?) (negative).
4. _Number of authors._
Only one author of both code and paper. Substantial input from others is not obvious. Correct? (negative).
5. _Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged._
The software consists of a single file `Hypercomplex.hpp`. That file has around 1100 lines of code. However, it is header only, so I guess that might be okay (positive).
6. _Whether the software has already been cited in academic papers._
I can not find any references (please correct me!) (negative).
7. _Whether the software is sufficiently useful that it is likely to be cited by your peer group._
Can not really judge on that one (inconclusive).

#openjournals/joss-reviews#3096
