### 2021-03-19 AngryMaciek/hypercomplex#26 @bonh
To summarize: I am not sure yet whether the contribution meets the requirements of JOSS for substantial scholarly effort.

Please, let us discuss this!
