### 2021-03-19 AngryMaciek/hypercomplex#26 @bonh
For example, I am wondering if it wouldn't be more suitable for your header file to be included into boost itself? If not, why?
