### 2021-03-19 #3096 @bonh
@melissawm I suggest that you get into that one #AngryMaciek/hypercomplex#26. Perhaps I am misreading the guidelines of JOSS.
