### 2021-03-24 AngryMaciek/hypercomplex#26 @melissawm
@AngryMaciek, do you feel like this is something you can respond to? Can you argue your case, so to speak? My feelings on the requirements are that if the usefulness of the software for future research use is established, and the software is sufficiently "complete" (as it appears to be agreed by all, including the comments on the review issue), it is reasonable to accept it. But I'd like to hear from the author.
