### 2021-03-24 AngryMaciek/hypercomplex#26 @AngryMaciek
Dear @bonh

Let me clarify some of your concerns:

The repository was initialised on 22 October 2020 which makes it a little over 5 months right now. I would claim that prior to starting the actual coding I dedicated some effort to investigating the subject and conceptualising the task but, as there will be no proof of that, you do not have to take it into account. And yes - I am the sole author of the project.

As for the number of commits and the commit history - this is more complicated, I believe.
In my personal opinion a good software engineering project will not have a long commit history on the main branch. Proper project management (either in [git-flow](https://nvie.com/posts/a-successful-git-branching-model/) or [GiHub flow](https://guides.github.com/introduction/flow/) model) is based on Pull Requests from feature branches. Each of such branches is responsible for a given modular functionality (or a bugfix). While merging them into the main branch it is advised to use the [_"Squash and merge"_](https://docs.gitlab.com/ee/user/project/merge_requests/squash_and_merge.html) option exactly in order to keep the commit history cleaner and more concise. Therefore the history of commits on the main branch would be _relatively_ short and transparent.
I learned these models only somewhere in the middle of my work on this project and as a result some of the commits have been pushed directly to the `master` branch whereas the rest are hidden in the [merged PRs](https://github.com/AngryMaciek/hypercomplex/pulls?q=is%3Apr+is%3Aclosed). It would be fair to take those into account too if we want to count them all (which I am not a fan of anyway, for the resons above).

I just took a closer look at 5 other [submissions which were recently accepted in _JOSS_](https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue+accepted+is%3Aclosed):
- https://github.com/openjournals/joss-reviews/issues/2968
- https://github.com/openjournals/joss-reviews/issues/3009
- https://github.com/openjournals/joss-reviews/issues/2957
- https://github.com/openjournals/joss-reviews/issues/3089
- https://github.com/openjournals/joss-reviews/issues/3049

Some of them have much shorter of a commit history, even **without** a feature branch-based model. Having this context I think that my repository meets the standards of a _JOSS_ publication in terms of this criterion.

As for the total lines of code: I would like to bring your attention to the fact that the repository itself is not only the core header file but also all the software engineering setup around it. Please take a look at how the [official _JOSS_ bot counts the lines](https://github.com/openjournals/joss-reviews/issues/3096#issuecomment-795850551). As you probably already noted - the count of _C/C++ Header_ lines is artificially higher since I include the [Catch2 test framework](https://github.com/catchorg/Catch2) in my repository (according to developers' instructions). My opinion is that, in line with the current bot setup, it is important to also consider other types of files (metadata, CI configuration, documentation, tests, development Makefile) in that statistics too, especially if main concernt here is an **open-source software**. These features should also count as my work as they directly indicate the state of the library as a whole (making it easier for the end users to utilise as well as for other developers to contribute to). If I remember correctly, _JOSS_ guidelines for authors also highlight the importantce of automatic tests and documentation for the open source community too (that is one of the reasons I dedicated a lot of effort into the software engineering part).

The software was not cited yet but I would not expect it to draw much attention in general. This is a highly specialised library for computational mathematicians and similar scientists/engineers. I do, however, believe that it may contribute to a progress in new directions, emerging in the fields of physics, robotics and deep learning (please take a look at https://github.com/AngryMaciek/hypercomplex/issues/25 where me and the other reviewer discussed potential applications). Relatively recent publications describe approaches which involve computations in higher-order algebras, but they are still quite novel. Lack of citations should not be much of a drawback at this stage, as (if I see this correctly) other _JOSS_ submissions I mentioned above were not cited prior to the acceptance too...

In regards to the nature of this library I also believe that (at least for now) it would be better to keep this code separate instead of bringing it under _boost_ from the start. I see that [Asio](http://think-async.com/Asio/) is an example of such library: started independently but later included into _boost_ and actively developed in both versions. This might be a matter of taste but I feel that the collection of _boost_ libraries are aimed at more specific, low-level tasks, whereas `Hypercomplex` is really dedicated to a smaller audience of specialists. I have checked what other libraries are there in the collection and I do not feel like my projects fits well among them... I did notice that the [Math](https://www.boost.org/doc/libs/1_75_0/libs/math/doc/html/index.html) library in _boost_ have their own definition of Quaternions and Octonions but merging my work right in there might not be as easy as I would like it to... On one hand I would not want to clash with those but on the other - the abstraction of all Cayley-Dickson construction will not allow me to specify as many constructors, operators and specialisations as for the _boost_ objects. Since they have the dimension of the algebra fixed for these two **separate** classes it is easier to provide a rich interface to the end user, unlike in my case where generality is the key and primary advantage.
Additionally, I would also like to keep the Apache 2.0 License for my work instead of publishing it with the Boost Software License (for patent specifications [described here](https://opensource.stackexchange.com/questions/7964/why-is-the-apache-license-2-0-patent-license-clause-useful-important)).
If, however, in the future I decide to bring this project into the _boost_ collection this should be a separate milestone in the project, in my opinion.

I hope these information shed some more light on the raised issues.

---
@melissawm: I was just constructing my reply today when your comment appeared a few minutes ago, sorry it took so long.
