### 2021-03-25 AngryMaciek/hypercomplex#26 @bonh
Hi @AngryMaciek,

thank you so mich for engaging in this discussion. Please note, that I am just trying to follow JOSS guidelines. My own opinion will come afterwards. Below I will question some of your arguments for the sake of JOSS (not necessarily my opinion!). @melissawm has to make the final decision here!

> I just took a closer look at 5 other [submissions which were recently accepted in _JOSS_](https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue+accepted+is%3Aclosed):
>
> * [openjournals/joss-reviews#2968](https://github.com/openjournals/joss-reviews/issues/2968)
> * [openjournals/joss-reviews#3009](https://github.com/openjournals/joss-reviews/issues/3009)
> * [openjournals/joss-reviews#2957](https://github.com/openjournals/joss-reviews/issues/2957)
> * [openjournals/joss-reviews#3089](https://github.com/openjournals/joss-reviews/issues/3089)
> * [openjournals/joss-reviews#3049](https://github.com/openjournals/joss-reviews/issues/3049)
>
> Some of them have much shorter of a commit history, even **without** a feature branch-based model. Having this context I think that my repository meets the standards of a _JOSS_ publication in terms of this criterion.

As far as I know I did not participate as a reviewer for the five publications you mention. And I do not want to (and cannot!) discuss the decision the editors did back then. However, I will not base my own discussion on previously published material in JOSS but solely my own opinion strongly guided by the guidelines published by JOSS! In my opinion, a future JOSS publication does not merely have to meet the quality of a subset of past publications in JOSS.

> As for the total lines of code: I would like to bring your attention to the fact that the repository itself is not only the core header file but also all the software engineering setup around it. Please take a look at how the [official _JOSS_ bot counts the lines](https://github.com/openjournals/joss-reviews/issues/3096#issuecomment-795850551). As you probably already noted - the count of _C/C++ Header_ lines is artificially higher since I include the [Catch2 test framework](https://github.com/catchorg/Catch2) in my repository (according to developers' instructions). My opinion is that, in line with the current bot setup, it is important to also consider other types of files (metadata, CI configuration, documentation, tests, development Makefile) in that statistics too, especially if main concernt here is an **open-source software**. These features should also count as my work as they directly indicate the state of the library as a whole (making it easier for the end users to utilise as well as for other developers to contribute to). If I remember correctly, _JOSS_ guidelines for authors also highlight the importantce of automatic tests and documentation for the open source community too (that is one of the reasons I dedicated a lot of effort into the software engineering part).

Good point! However, I understand the guidelines in a different way (counting only core code lines, test and documentation is mandatory for acceptance in JOSS), but I might be wrong - @melissawm, please clarify!

> The software was not cited yet but I would not expect it to draw much attention in general. This is a highly specialised library for computational mathematicians and similar scientists/engineers. I do, however, believe that it may contribute to a progress in new directions, emerging in the fields of physics, robotics and deep learning (please take a look at #25 where me and the other reviewer discussed potential applications). Relatively recent publications describe approaches which involve computations in higher-order algebras, but they are still quite novel. Lack of citations should not be much of a drawback at this stage, as (if I see this correctly) other _JOSS_ submissions I mentioned above were not cited prior to the acceptance too...

Unfortunately, I did not get your discussion with the other reviewer. Was is referenced inside the JOSS review issue?
Still, your software has never been used in research works, right? In terms of JOSS guidelines this might hinder publication, @melissawm?

> In regards to the nature of this library I also believe that (at least for now) it would be better to keep this code separate instead of bringing it under _boost_ from the start. I see that [Asio](http://think-async.com/Asio/) is an example of such library: started independently but later included into _boost_ and actively developed in both versions. This might be a matter of taste but I feel that the collection of _boost_ libraries are aimed at more general, low-level tasks, whereas `Hypercomplex` is really dedicated to a smaller audience of specialists. I have checked what other libraries are there in the collection and I do not feel like my projects fits well among them... I did notice that the [Math](https://www.boost.org/doc/libs/1_75_0/libs/math/doc/html/index.html) library in _boost_ have their own definition of Quaternions and Octonions but merging my work right in there might not be as easy as I would like it to... On one hand I would not want to clash with those but on the other - the abstraction of all Cayley-Dickson construction will not allow me to specify as many constructors, operators and specialisations as for the _boost_ objects. Since they have the dimension of the algebra fixed for these two **separate** classes it is easier to provide a rich interface to the end user, unlike in my case where generality is the key and primary advantage.
> Additionally, I would also like to keep the Apache 2.0 License for my work instead of publishing it with the Boost Software License (for patent specifications [described here](https://opensource.stackexchange.com/questions/7964/why-is-the-apache-license-2-0-patent-license-clause-useful-important)).
> If, however, in the future I decide to bring this project into the _boost_ collection this should be a separate milestone in the project, in my opinion.

In my opinion, based on the guidelines, this speaks against a publication in JOSS _right now_. It sounds, as if you had a great idea, came up with a really nice implementation but now your software package is lacking application and experience. Perhaps you might work out an application of your software and publish a research paper on it? If a few people start using it a re-implementation in boost might follow. @melissawm, I am lost here and need a final decision from you whether the scholarly effort in the current state validates publication in JOSS! Note, that if you decide in favor of @AngryMaciek I will definitely want to continue my review, right now it is just about the interpretation of JOSS guidelines.
