### 2021-03-25 AngryMaciek/hypercomplex#26 @bonh
Sry, my fault, the issue by the other reviewer was referenced in the review issue!
