### 2021-03-29 AngryMaciek/hypercomplex#26 @melissawm
Hello, @bonh and @AngryMaciek !

I'll cite one piece of the Review Guidelines from JOSS:

> These guidelines are not meant to be strictly prescriptive. Recently released software may not have been around long enough to gather citations in academic literature. While some authors contribute openly and accrue a long and rich commit history before submitting, others may upload their software to GitHub shortly before submitting their JOSS paper. Reviewers should rely on their expert understanding of their domain to judge whether the software is of broad interest (likely to be cited by other researchers) or more narrowly focused around the needs of an individual researcher or lab.

So, in my view, those are not requirements. However, we will respect the reviewer's concerns if they think this submission should be flagged for a scope check. Is that the case, @bonh ? If so, I will share this with other JOSS editors and we can discuss our findings and get back to you. Thanks!
