### 2021-03-29 AngryMaciek/hypercomplex#26 @bonh
Thank you @melissawm for clarification!

My "expert understanding" of everything above Octonions is basically zero. So I googled for applications of "Sedenions" but @AngryMaciek already stated that in #25 that there seems to be only a "single resource" (which is not a concern per se!). So I am completely unfit

> to judge whether the software is of broad interest (likely to be cited by other researchers).

My gut feeling is that it is premature to publish a paper on the library in JOSS right now. So based on the note right below the piece of the Review Guidelines from JOSS reference by @melissawm above:

> The decision on scholarly effort is ultimately one made by JOSS editors. Reviewers are asked to flag submissions of questionable scope during the review process so that the editor can bring this to the attention of the JOSS editorial team.

I flag this submission for scope check.
