### 2021-03-29 #3096 @bonh
@melissawm based on #AngryMaciek/hypercomplex#26 I flag this submission for scope check.
