### 2021-03-29 #3096 @melissawm
@bonh @AngryMaciek based on discussions with the JOSS editors, since one reviewer has expressed approval of the submission, we feel the best course of action would be to invite a third reviewer with more familiarity with this field in order to complete the review. In that case, @bonh would be removed as a reviewer. Let me know if you agree.
