### 2021-03-29 #3096 @AngryMaciek
I am perfectly fine with such resolution.
