### 2021-04-09 #3096 @melissawm
Hello, @iljah! Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

You can check a bit of the history of this review here in the thread, let me know if you have any questions or need more information.
