### 2021-04-12 #3096 @iljah
@melissawm thanks for the opportunity but I have very little experience even with regular complex numbers and no experience with more complex numbers. @AngryMaciek what did stand out from documentation (https://angrymaciek.github.io/hypercomplex/html/index.html) was need for manual memory management when using MPFR types. I've almost not used MPFR but at least it should be possible to automate much of what is done manually at end of above documentation. Following notice is also unusual for header-only c++ library: "Please do not mix this library with another MPFR-dependant code in the same translation unit.".
