### 2021-04-14 #3096 @AngryMaciek
Thanks for the comment @iljah!
1.  Each _MPFR_  variable has to be manually freed from the memory, that's how the library is designed. In the specialisation for `Hypercomplex` which is based on these types I do free the internal variables in the [class destructror](https://github.com/AngryMaciek/hypercomplex/blob/2f7db6fdc4dbeff9b5115a8244f4d7c8819b162f/hypercomplex/Hypercomplex.hpp#L652). These commands you probably refer to, at the end of the documentation, are related to freeing memory from an (mock) array which user could create in their main program.
2. 'Non-mixing' information is there because of how _MPFR_ stores precision related settings in global variables, but maybe the notice is obsolete now. I will have to check this once again in the documentation...
