### 2021-04-15 #3096 @iljah
I mean that at least code shown in documentation could be moved to hypercomplex constructor I think. Hypercomplex<mpfr_t, 8> already knows that there will be 8 mpfr types so it could call mpfr_init2 in constructor. Constructor could also call mpfr_set_d which would allow using H{1,2,3,...} directly. Although now I notice that same thing applies to doubles so hypercomplex never owns its storage? That's surprising since all sizes are known at compile time.
