### 2021-04-16 #3096 @AngryMaciek
@iljah OK thanks for the input! I remember that there were some constraints on how the constructor would work due to the recursive implementation of the multiplication operator. I will for sure look into this deeper once we move on with the review.
