### 2021-04-19 #3096 @melissawm
Hello, @nnadeau - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

You can check a bit of the history of this review here in the thread. Let me know if you have any questions or need more information.
