### 2021-04-20 #3096 @nnadeau
@melissawm sure, i can be available. "hypercomplex numbers" are outside of my direct specialty, but I can definitely review the C++ and the overall scholarly effort.
