### 2021-04-22 #3096 @nnadeau
@melissawm @AngryMaciek , this is a great tool, but I cannot recommend publication based on the grounds of [`Substantial scholarly effort`](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

- Only represents about 2 months of work ([judging by commit history](https://github.com/AngryMaciek/hypercomplex/graphs/contributors?from=2020-10-18&to=2020-12-31&type=c))
- Only <1000 lines of code, much of which is just C++ templating

```
~/git/hypercomplex ‹master› » cloc hypercomplex                                       2 ↵
       2 text files.
       2 unique files.
       0 files ignored.

Language                     files          blank        comment           code
C/C++ Header                     1             95            338            668
make                             1             17             31             35
SUM:                             2            112            369            703
```

- The `Statement of need` does not convince me **why** I should choose to use this library vs simply writing my own small utility functions to handle this use case. Including 3rd party packages is a risk and there has to be a significant value proposition to justify the risk.
- Yes, Boost is bloated and doesn't include this functionality, but this tool would make the most positive impact on the world if it was contributed to the Boost project
