### 2021-04-29 #3096 @AngryMaciek
Dear @melissawm,

Yes, of course! I would very much like to respond to the points raised by @nnadeau:

[1] I am not sure why, but that contribution graphs do not reflect the commit history accurately. Please take a look at the [commit history](https://github.com/AngryMaciek/hypercomplex/commits/master) of `master` directly. Initial commit for this repo dates back to 22 October, the last commit is from 16 February. That would result in not 2 months but official 3 months and 3 weeks of work (so, almost 2x increase). Of course it's not only the number of days that matters, but I am just replying to the reviewers comment. According to the explicit specifications in the [JOSS Submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), this should be fine:
> As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual.

[2] The statistics which the reviewer posted above are correct, the core template file is not big. However, I believe that we should stick to counting the lines of code just as _JOSS_ does, i.e. please look at [this post](https://github.com/openjournals/joss-reviews/issues/3096#issuecomment-795850551) by @whedon. `cloc` is configured here to explicitly count formats such as XML, YAML, Markdown, HTML etc. These are not pure code but their reflect my work and contribute to the overall high state of the library from the software engineering perspective (CI, documentation, READMEs, tests). Personally, it is not a surprise for me that @whedon is configured here in such manner - after all this journal emphasizes the high-quality software engineering therefore it seems reasonable that it counts files other then the core code.
Please note that inside this repository I also include [Catch2](https://github.com/catchorg/Catch2) testing library, so we need to deduct it from the `cloc` output:
```
$ cloc hypercomplex/.test/unit/catch.hpp
       1 text file.
       1 unique file.
       0 files ignored.

Language                     files          blank        comment           code
C/C++ Header                     1           3138           1070          13617
```
Therefore, in the end we are left with `16224 - 13617 = 2607` lines of code for the whole repository, as counting by the _JOSS_ standards. Again, quoting the submission instructions: such number should not raise concerns regarding the repository size:
> Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.

[3] I cannot really say I agree with this mindset. We are having here a _Journal of Open Source Software_ which is - as it states - mostly focused on the quality of the software, CI/CD testing, documentation etc. This is exactly so that not everyone would need to write custom functions (and their unit tests!) from scratch. This is one of the main ideas behind open source software - so that a large group of people share small building blocks. In that way we can collectively build bigger and greater things. Modularisation is a key feature in the process.
Additionally, about the risk - I claim that it is quite the opposite. Including a high-quality library with its own testing framework might be less of a risk than relying on a custom piece of code which was written on-the-fly.
I cannot speak for others than myself but generally when I see a solid library out there I am happy to use it instead of reinventing the wheel myself. And the submission guidelines for _JOSS_ enforce that the libraries published here *are* solid (so that other people can indeed rely on my work).

[4] OK, I agree here. If you believe it makes sense I will reach out to _Boost_ developers and ask if they would be interested in including _Hypercomplex_(?)
