### 2021-06-20 #3096 @bonh
Hi @melissawm,
the submission is still flagged for editorial review - so it is not done yet and I feel uncomfortable to start another review while that one still lingers around (even if I'm not a part of it anymore). My point is, that I might not have the heart to ask for another scope check (or some other editorial stuff) during the review of another submission even if I feel it might be needed - after all we all do this on the side and I greatly appreciate the work of the editors! I hope you get what I mean and please let me know if I'm stepping on someone's toes here (if so I am sorry!).
