### 2021-06-20 #3096 @danielskatz
@melissawm - to flag the AiEC, please mention @openjournals/joss-eics Doing a query scope command is generally intended for pre-review and doesn't get tracked well when done by an editor during a review.
