### 2021-10-01 #3096 @AngryMaciek
@danielskatz
Yes, sorry for the delay. It took me much longer then expected since I needed to learn more theory about advanced cryptography, before even having started to code it. I am essentially adding here a class specialisation to include the functionality described [in this paper](https://eprint.iacr.org/2009/386.pdf) I think it fits perfectly as a concrete application of a general algebraic framework presented in this repo.
I have a strong deadline at work: until the end of this month I am super busy with everyday tasks. I will be able to come back here and finish this off in November.
