### 2021-11-30 #3096 @AngryMaciek
Dear @danielskatz and @melissawm,
I still need more time to implement the changes I envison for this library.
Maybe it would be better then if I resubmit to _JOSS_ once everything is finished?
